General refactoring
- adding configuration and image manipulation classes
- Separating configuration on proper enums
